feat: replace DOMAIN with assets path
This commit replaces the usage of the DOMAIN environment variable with
the assets path provided by SvelteKit. This change affects the favicon
URL in the layout and the domain URL in the feed server. Additionally,
the svelte.config.js file is updated to set the assets path based on the
DOMAIN environment variable or a default value.